corrected logic for fonts from idea at mac

+++ modules/base/editor-ui-api/src/main/java/com/intellij/ide/ui/UISettings.java
@@ -459,7 +459,7 @@ public class UISettings extends SimpleModificationTracker implements PersistentS
     RIGHT_HORIZONTAL_SPLIT = value;
   }
 
-  public float getFontSize() {
+  public int getFontSize() {
     return FONT_SIZE;
   }
